tools/requirements.txt
PyYaml
click
six
python-dateutil
python-slugify